+++ src/frontend/apps/drive/src/features/drivers/Driver.ts
@@ -1,5 +1,6 @@
 import { ExplorerFilterModifiedValue } from "../explorer/components/filters/ExplorerFilterModified";
 import {
+  DTOBatchShare,
   DTOCreateAccess,
   DTODeleteAccess,
   DTOUpdateAccess,
@@ -142,6 +143,7 @@ export abstract class Driver {
   abstract deleteFavoriteItem(itemId: string): Promise<void>;
   abstract getItemAccesses(itemId: string): Promise<Access[]>;
   abstract createAccess(data: DTOCreateAccess): Promise<void>;
+  abstract batchShare(payload: DTOBatchShare): Promise<void>;
   abstract updateAccess(payload: DTOUpdateAccess): Promise<Access | void>;
   abstract updateLinkConfiguration(
     payload: DTOUpdateLinkConfiguration,